# docs/swagger.yaml
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      username:
        type: string
    type: object
  auth.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.ResumeResponse:
    properties:
      best_times:
        items:
          $ref: '#/definitions/auth.gameResult'
        type: array
      today_games:
        items:
          $ref: '#/definitions/auth.gameResult'
        type: array
      total_games:
        additionalProperties:
          type: integer
        type: object
    type: object
  auth.gameResult:
    properties:
      finished:
        type: boolean
      size:
        type: integer
      time:
        type: integer
    type: object
  leaderboard.Entry:
    properties:
      rank:
        type: integer
      username:
        type: string
      value:
        type: string
    type: object
  leaderboard.GetLeaderBoardResponse:
    properties:
      has_next:
        type: boolean
      solves:
        items:
          $ref: '#/definitions/leaderboard.Entry'
        type: array
    type: object
  leaderboard.ResetStrikesRequest:
    properties:
      date:
        type: string
    type: object
  pkg.Error:
    properties:
      code:
        type: string
      message:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/pkg.ValidationError'
        type: array
    type: object
  pkg.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  src_infrastructure_http_sudoku.Solve:
    properties:
      date:
        type: string
      duration:
        type: integer
      id:
        type: string
      size:
        type: integer
      started_at:
        type: string
    type: object
  sudoku.Cell:
    properties:
      col:
        type: integer
      row:
        type: integer
      value:
        type: integer
    type: object
  sudoku.GetDailySudokuResponse:
    properties:
      board:
        items:
          $ref: '#/definitions/sudoku.Cell'
        type: array
      date:
        type: string
      id:
        type: string
      session_id:
        type: string
      session_token:
        type: string
      size:
        type: integer
      started_at:
        type: string
    type: object
  sudoku.MySolvesResponse:
    properties:
      solves:
        items:
          $ref: '#/definitions/src_infrastructure_http_sudoku.Solve'
        type: array
    type: object
  sudoku.VerifySolutionRequest:
    properties:
      play_token:
        type: string
      solution:
        items:
          items:
            type: integer
          type: array
        type: array
    required:
    - play_token
    - solution
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access token. Refresh token is
        sent via secure HTTP-only cookie.
      parameters:
      - description: User login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: invalid_body
          schema:
            $ref: '#/definitions/pkg.Error'
        "401":
          description: invalid_credentials
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Login user
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the user's refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            type: string
        "401":
          description: invalid_token
          schema:
            $ref: '#/definitions/pkg.Error'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes an expired access token using refresh token from cookie.
        Returns new access token and rotates refresh token cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshTokenResponse'
        "401":
          description: invalid_token, refresh_token_expired, refresh_token_revoked
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Refresh access token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
        "400":
          description: invalid_body, validation_error
          schema:
            $ref: '#/definitions/pkg.Error'
        "409":
          description: email_already_registered
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Register a new user
      tags:
      - auth
  /api/auth/resume:
    get:
      description: Returns user statistics including total games, today's games, and
        best times
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ResumeResponse'
      security:
      - BearerAuth: []
      summary: Get user resume
      tags:
      - auth
  /api/cron/unfinished-attempts:
    post:
      consumes:
      - application/json
      description: Removes unfinished attempts for the daily sudoku puzzles that are
        past the reset threshold (e.g., 24 hours). This endpoint is intended to be
        called by a scheduled job to clean up old attempts and reset strikes for users
        who haven't completed their puzzles in time.
      produces:
      - application/json
      responses:
        "200":
          description: Cleaned unfinished attempts
      summary: Remove unfinished attempts
      tags:
      - sudoku
  /api/leaderboard:
    get:
      consumes:
      - application/json
      description: Returns the leaderboard with rankings for a given type and size.
        For daily and all-time types, size is required. For streak and total types,
        size should not be provided.
      parameters:
      - description: Leaderboard type (daily, all-time, streak, total)
        in: query
        name: type
        required: true
        type: string
      - description: Board size (four, six, nine) - required for daily and all-time,
          not allowed for streak and total
        in: query
        name: size
        type: string
      - description: Number of entries to return (1-100)
        in: query
        name: limit
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/leaderboard.GetLeaderBoardResponse'
        "400":
          description: invalid_leaderboard_type, invalid_size, invalid_limit, invalid_page,
            size_required, size_not_allowed
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Get leaderboard
      tags:
      - leaderboard
  /api/leaderboard/reset:
    post:
      consumes:
      - application/json
      description: Resets current_streak to 0 for users whose last_solved_date is
        before yesterday. Called daily by Cloud Scheduler. If no date is provided,
        defaults to the current time.
      parameters:
      - description: Date threshold for reset (optional, defaults to now)
        in: body
        name: request
        schema:
          $ref: '#/definitions/leaderboard.ResetStrikesRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content - strikes reset successfully
        "400":
          description: invalid_body, validation_error
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Reset leaderboard strikes
      tags:
      - leaderboard
  /api/sudoku:
    get:
      consumes:
      - application/json
      description: Returns the daily sudoku puzzle for a given size
      parameters:
      - enum:
        - four
        - six
        - nine
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudoku.GetDailySudokuResponse'
        "400":
          description: invalid_query_param, invalid_size
          schema:
            $ref: '#/definitions/pkg.Error'
        "404":
          description: sudoku_not_found
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Get daily sudoku
      tags:
      - sudoku
  /api/sudoku/generate/{size}:
    post:
      consumes:
      - application/json
      description: Generates new daily sudoku puzzle for a given size
      parameters:
      - description: Board size (four, six, or nine)
        enum:
        - four
        - six
        - nine
        in: path
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudoku.GetDailySudokuResponse'
        "400":
          description: invalid_size
          schema:
            $ref: '#/definitions/pkg.Error'
      summary: Generate sudoku
      tags:
      - sudoku
  /api/sudoku/me:
    get:
      description: Returns the daily sudoku solves for a given user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudoku.MySolvesResponse'
        "401":
          description: invalid_token
          schema:
            $ref: '#/definitions/pkg.Error'
      security:
      - BearerAuth: []
      summary: Get my daily sudoku solves
      tags:
      - sudoku
  /api/sudoku/submit:
    post:
      consumes:
      - application/json
      description: Verifies if the submitted solution is correct
      parameters:
      - description: Solution request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudoku.VerifySolutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Solution verified successfully
          schema:
            type: string
        "400":
          description: invalid_body, invalid_solution
          schema:
            $ref: '#/definitions/pkg.Error'
        "401":
          description: invalid_token
          schema:
            $ref: '#/definitions/pkg.Error'
        "404":
          description: solution_not_found
          schema:
            $ref: '#/definitions/pkg.Error'
        "409":
          description: already_played
          schema:
            $ref: '#/definitions/pkg.Error'
      security:
      - BearerAuth: []
      summary: Verify sudoku solution
      tags:
      - sudoku
  /api/sudoku/submit/guest:
    post:
      consumes:
      - application/json
      description: Verifies if the submitted solution is correct for guest users
      parameters:
      - description: Solution request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudoku.VerifySolutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Solution verified successfully
          schema:
            type: string
        "400":
          description: invalid_body, invalid_solution
          schema:
            $ref: '#/definitions/pkg.Error'
        "401":
          description: invalid_token
          schema:
            $ref: '#/definitions/pkg.Error'
      security:
      - BearerAuth: []
      summary: Verify guest sudoku solution
      tags:
      - sudoku
swagger: "2.0"